image-captioning/requirements.txt
Flask
google-generativeai
google-cloud-texttospeech
Pillow
python-dotenv